Reduce boilerplate when dialing with timeout
